scripts/sidecar/requirements.txt

boto3==1.34.0
requests==2.31.0
Pillow==10.2.0
